Improved install. Separated aliases to partials

+++ install.sh
@@ -1,20 +1,23 @@
 #!/usr/bin/env bash
 
-echo "Checking if Oh My ZSH is installed..."
+read -p "This will delete your custom aliases if you had them. Are you sure [y/N]?" -n 1 -r
+echo
+echo
 
-if [ -d ~/.oh-my-zsh ]; then
-    echo "Oh My ZSH is installed."
-    echo "Linking custom aliases..."
-    
-    echo "Checking if aliases already exist..."
-    if [ -a ~/.oh-my-zsh/custom/aliases.zsh ]; then
-        rm ~/.oh-my-zsh/custom/aliases.zsh
-        ln -s $PWD/zsh/aliases.zsh ~/.oh-my-zsh/custom/aliases.zsh
-        echo "Aliases are linked successfully."
-    else
-        ln -s $PWD/zsh/aliases.zsh ~/.oh-my-zsh/custom/aliases.zsh
-        echo "Aliases are linked successfully."
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+    if [ -d ~/.oh-my-zsh ]; then
+        if [ -a ~/.oh-my-zsh/custom/aliases.zsh ]; then
+            rm ~/.oh-my-zsh/custom/aliases.zsh
+            cat $PWD/zsh/aliases/*.zsh > ~/.oh-my-zsh/custom/aliases.zsh
+            echo "Aliases was created successfully!"
+        else
+            cat $PWD/zsh/aliases/*.zsh > ~/.oh-my-zsh/custom/aliases.zsh
+            echo "Aliases was created successfully!"
+        fi
+    else 
+        echo "Please install Oh My ZSH."
     fi
-else 
-    echo "Please install Oh My ZSH."
+else
+    echo "Aborted!"
 fi
